**BREAKING CHANGES** - Refactoring of commands   - Dispatch struct is no more. All attached struct are now located at `https://github.com/edjubert/hyprland-ipc-go/hyprctl/dispatch`:     - Move     - Focus     - Toggle     - Resize   - `Get` struct is now located under `https://github.com/edjubert/hyprland-ipc-go/hyprctl/get`   - `SendNotification` is now located under `https://github.com/edjubert/hyprland-ipc-go/hyprctl/notify`   - Socket function `Create` and `Remove` are now under `https://github.com/edjubert/hyprland-ipc-go/socket` - New function to write directly on the hyprland socket called `WriteCmd`
